TalkitOut-Backend (REST-APIs)/chatbot/views.py
from rest_framework.decorators import api_view
from rest_framework.response import Response
from rest_framework import status


@api_view(["POST"])
def chatbotResponse(request):
    if request.method == "POST":
        data = request.data
        print(data)
        return Response("Hello!", status=status.HTTP_200_OK)
    else:
        return Response(
            {"message": "Method not allowed"}, status=status.HTTP_405_METHOD_NOT_ALLOWED
        )
