requirements.txt
asgiref==3.7.2
channels==4.0.0
Django==5.0
django-crispy-forms==2.1
django_ckeditor==6.7.0
django_richtextfield==1.6.1
djangorestframework==3.14.0
Pillow==10.1.0
Pillow==10.1.0
